Added brick object constuctor, code cleanup, and bug fix
-Added constructor for brick objects.
-Cleaned up code in createBricks() to use new constructor
-Updated brick colision dectection functions to use targetBrickIndex.
Brick objects no longer need to track their index in the array.
-Fixed bug in collideBrick() where left and right colisions were never
checked. This introduced a new error where the ball can go between
bricks if the ball's x velocity is 0, Resolved this error by setting
starting x velocity to 1.

Known issue:
The ball can hit the same brick multiple times when it hits the corner
of a brick.